Bugfix + safe fallback if WebSpeech isn't ready yet
The getVoices() method of window.speechSynthesizer prompts the WebSpeech
API to asynchronously prepare a list of available voices. This normally
takes an insignificant amount of time, however, to play it safe, I have
introduced a change whereby the application defaults to using the
<audio> elements at first, then checks if the WebSpeech API is
available, and if so, asks it to prepare the voices list. In case this
took long, we make sure that until the list is ready, we're still using
the <audio> elements. Only once the list is ready and the API has
dispatched a "voiceschanged" event do we assign a different handler for
the click events, which this time plays synthesized speech instead of
pre-rendered recordings.
The previous two commits introduced an error whereby the
"removeDefaultVoice" (later "init") function did not check if the
WebSpeech API is available. The present commit no longer has that error
as we need to perform only one check in the entire application.