Refactoring stage 5
So far, we invoked the "getVoices" method on every click, which - even
though it's not very resource-consuming - is rather inefficient. Now
I've created a global (i.e. global to the app's scope) variable "voice".
Its value doesn't matter before we've switched to WebSpeech but once we
do, we call getVoices for the second (and last) time and assign its
fifth value to the "voice" variable. Then our "playSynth" function can
access that variable without having to query the API every time. Other
changes:
- Brought back the "initSynth" function since there are now two
operations we need to perform, so it's better to extract them to a
separate function.
- Renamed "color" to "utterance" because it's an utterance and not a
color and actually, notice now that the "playSynth" function has nothing
to do with colors - you could use the same exact code for any other type
of content without changing anything.