Refactoring stage 6
Notice how the "voice" variable was unnecessarily global. The playAudio
function had access to it even though it was not concerened with it at
all. This was not good. In this commit I've encapsulated the "play*"
functions into player objects and I've put the previously global
"player" variable in another scope so the players can't see or modify
it, because why would they - players are supposed to play sound, not
change application state (in this case, which of the players the app
decides to use). So now the app initially decides to use "audio" as its
player - any time there is a "click" event, it's "play" method will be
called with the event object passed as the argument. Then we check if
WebSpeech is available, and if so, we initialize the "synth" player.
Since this initialization is asynchronous (we start the initialization
by calling "getVoices" once and then we wait an undefinite amount of
time till the API dispatches a "voiceschanged" event), we it would be
unsafe to write:
synth.init();
player = synth;
- because by the time we say "player = synth" the initialization hasn't
finished yet. So instead, we pass the "init" method a function which it
will call once the initialization has completed. Now the synth object
doesn't make itself the app's player, the app does it for itself in the
callback.