Changed "play" methods back to "handleClick"
At the moment these methods violate the Single Responsibility principle
because they deal both with the DOM (and obtain the content for
themselves) and the actual playing part. Eventually, this probably
should be split into separate functions but for now I've just renamed
them because I think "handleClick" is more general than "play" and
these methods actually are quite general (and they do take a click event
object as an argument instead something to play).